Use int64 as price_1000x
Update booker tests for new market order logic.

+++ include/libbooker/BookerCommonData.h
@@ -14,8 +14,8 @@ class BookerCommonData
 public:
     [[nodiscard]] static char to_side(types::SideType side);
     [[nodiscard]] static types::SideType to_side(char side);
-    [[nodiscard]] static liquibook::book::Price to_price(double price);
-    [[nodiscard]] static double to_price(liquibook::book::Price price);
+    [[nodiscard]] static liquibook::book::Price to_price(int64_t price);
+    [[nodiscard]] static int64_t to_price(liquibook::book::Price price);
     [[nodiscard]] static liquibook::book::Quantity to_quantity(int64_t quantity);
     [[nodiscard]] static int64_t to_quantity(liquibook::book::Quantity quantity);
